Cambio de repositorio

+++ vite-project/src/components/ItemDetailContainer/ItemDetailContainer.jsx
@@ -1,30 +1,46 @@
-import { useEffect, useState } from 'react';
-import ItemDetail from "./ItemDetail"
+import React, { useEffect, useState } from 'react';
+import ItemDetail from './ItemDetail';
 import { useParams } from 'react-router-dom';
-import {doc, getDoc} from "firebase/firestore";
+import { doc, getDoc } from 'firebase/firestore';
 import { db } from '../../firebase/config';
 
- const ItemDetailContainer = () => {
-    const [item, setItem] = useState(null);
-    //nos devuelve un objeto con los parametros que haya en url
-    const { id } = useParams();
+const ItemDetailContainer = () => {
+  const [item, setItem] = useState(null);
+  const { id } = useParams();
 
-    useEffect(() => {
-    
-      const docRef = doc (db, "productos", id)
-      getDoc (docRef)
-      .then ((resp)=>{
-        setItem  (
-          {...resp.data (), id: resp.id}
-        );
-      })
+  useEffect(() => {
+    const fetchItem = async () => {
+      try {
+        if (!id) {
+          throw new Error('ID no válido');
+        }
 
-    }, [id])
-    
+        const docRef = doc(db, 'productos', id);
+        const resp = await getDoc(docRef);
+
+        if (resp.exists()) {
+          setItem({
+            ...resp.data(),
+            id: resp.id,
+          });
+        } else {
+          console.log('Documento no encontrado');
+          setItem(null); 
+        }
+      } catch (error) {
+        console.error('Error al obtener el documento:', error);
+        setItem(null);
+      }
+    };
+
+    fetchItem();
+  }, [id]);
+
   return (
     <div>
-        {item && <ItemDetail item={item} />}
+      {item ? <ItemDetail item={item} /> : <p>Item no encontrado</p>}
     </div>
-  )
-}
+  );
+};
+
 export default ItemDetailContainer;
